% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbkdf.R
\name{pbkdf}
\alias{pbkdf}
\title{Bcrypt PWKDF}
\usage{
pbkdf(password, salt, rounds = 16L, size = 32L)
}
\arguments{
\item{password}{string or raw vector with password}

\item{salt}{raw vector with (usually 16) bytes}

\item{rounds}{number of hashing rounds}

\item{size}{desired length of the output key}
}
\description{
Password based key derivation function with bcrypt.
}
